fix crawler bugs
When building the Request struct directly, http.Do would fail with an
invalid path variable, somehow the rawQuery would get mixed with the
path. Instead I switched to http.NewRequest instead, which is much
simpler anyway.

There was also a bug in the throttling mechanism; I originally forgot
to Sleep. Note that the wait duration calculation does not check for
negative numbers; that is because Go's Sleep will simply ignore
negative values silently anyway.

This commit also includes a safeguard mechanism to stop the crawling
at 10 modules, this is to avoid flooding the cdn while the crawling
logic is still being ironed out.