add better instrumentation for crawler
This commit includes refactors to the crawler into it's own file and
into a more defined struct and interface. It uses the promhttp
package to expose various metrics from the underlying http client like
latencies, dns and tls overhead and number of http requests.

It also includes a fix to the prometheus and grafana stack to allow
scraping metrics from the Go process running on the host. The prometheus
and grafana instances are now running in the 'host' network, I've had
to adjust some of the configurations to adjust host names for that.