add inserts for Module and ModuleVersion entities
Adds a step to insert Module and ModuleVersion nodes into the graph.
I'm still unsure if I will keep those node types as-is, but this allowed
me to refactor some of the channel sync logic and overall flow of the
crawler and also fix one or two hidden bugs where channels weren't being
closed properly.