add DynamoDB table operations to crawler
This commit hooks DynamoDB to the mix; DynamoDB acts as a sort of ledger for
every file that goes into the Dgraph cluster. Because Dgraph isn't great at
querying for nodes using a specific key (in a "get me _this_ item kind of way),
using DynamoDB makes that much easier. It also replaces the in-memory that was
used for that purpose. The map worked well for a small number of modules or
files but the number of files increases quadratically with the number of
modules and is simply not an option at scale -- DynamoDB offloads this load
from memory to another computer.