add SQS delete message & fix memory leak
I've known about this memory leak for quite some time now but was never
able to put the finger on it. Thanks to @danopia I was able to figure
out it was because I was `defer`ing the Discard operation in each
iteration of the loop in InsertFiles, and passing it a reference to the
transaction object, which meant the runtime was keeping each transaction
instance, and its associated mutations in memory forever.